Added a `Document Launcher' plugin:
- first start the plugin
- browse your filesystem with the Shell plugin
- once you found an interesting file, just say `gedit-launch <filename>'
  to open that file in gEdit (the filename can be relative to the
  current directory).

Technically, this is implemented with a CORBA server which is started
in the gmodule plugin. The `gedit-launch' program then connects to this
CORBA server to open the document.

This is still a little bit experimental, but it shows you how we can
export the functionality of gEdit through CORBA without the need to
implement the CORBA interface in gEdit directly (only the naming service
is started there and the `global_orb' and the `root_poa' are set up there)
but in a plugin.

Martin <martin@home-of-linux.org>