Add the sort plugin by Carlo Borreo <borreo@softhome.net>.
sigh. I DON'T like gedit's code. But it is stable so
i can't change it at this point unless we go thru another
stabilization period. I guess i am going to wait till i need
to port it to GTK 2.0. It was a good learning experience tho
and the code is a lot bettter than waht it was before.

2001-06-04  Chema Celorio  <chema@celorio.com>

	* sort/sort.c: added the sort plugin by Carlo Borreo <borreo@softhome.net>

2001-06-04  Chema Celorio  <chema@celorio.com>

	* file.c (gedit_file_open): if the file is not an URI, ask the user if he
	wants to create it.