Implemented fallback strategy for saving files when opening files read/write is not supported
This should fix a problem with file systems that do not support opening files for read/write 
(such as the fuse file system: curlftpfs). The strategy is to open the file write only when read/write
fails with ENOTSUP. This means that we will not be able to make a backup (and the user receives
a warning for this), but at least we can safe the file now.