Remember filechooser location also for remote files
With gnomevfs we special cased remote files and in that case we did not
store the current file chooser dir, since that triggered blocking
behavior when opening the file chooser. Lift that limitations in the
hope that gvfs works better. Also convert all related code to deal with
GFiles directly. Bug #549465.