Improved external tool process handling
No longer use shell. This ensures that when the tool is cancelled/terminated, there is
no child process left running. Bind STDERR to STDOUT if both should be used (this will ensure
more consistent output). Lastly, if killing the tool cleanly does not work, a second cancellation
will end the process with SIGKILL.