Only apply default values for modelines when strictly necessary
This solves the bug where modeline will always restore default values for all settings
there are modelines for by default (even when there is no modeline in the document).

Basicly, modelines will now only restore to default if the current setting of a value was
due to the previous modeline.