Rework the document input stream to use a mark
This is necessary even if the text is not modified during the read since
gtksourceview or plugins can interact with the buffer applying tags and
changing marks. The patch hopefully fixes bug #611757, which I cannot
reproduce anymore, though we've not been able to create a simple unit
test isolating the bug.