Explicitly close output stream with cancellable in cancelled state
This fixes an issue in GOutputStream that applies the atomic replace
save strategy. The stream moves the written file to the original file
when the stream is closed. However, there is no way currently to tell
the stream that the save should be aborted (there could be a
conversion error). The patch explicitly closes the output stream
in all these cases with a GCancellable in the cancelled state, causing
the output stream to close, but not move the file. This makes use
of an implementation detail in the local gio file stream and should be
properly fixed by adding the appropriate API in gio. Until then, at least
we prevent data corruption for now.