Disable libpeas dependency for now.
It is planned to use libpeas for the plugin support, but it is a bit
premature to add a strong dependency on it right now. Besides, having
such a dependency has to be announced publicly following Gnome terms...

Let's do without for now, as we don't use libpeas for real yet.