Autogenerate settings flags and use them.
Now that gsettings allow to autogenate flags and provides methods
to use them. Let's not do the flags conversion and use them directly.
See that the settings compiler compiles the settings by alphabetic order
so we need to change the settings file in order to get the flags built
first.